requirements.txt
carla==0.9.15
numpy
opencv-python
hypothesis
pytest
torch
torchvision
pandas
pillow
gymnasium
